requirements.txt
chromadb>=0.4.22
sentence-transformers>=2.2.2
gradio>=4.0.0
openai>=1.0.0
python-docx>=1.0.0
odfpy>=1.4.1
